First commit to have atleast the bare basic files
First commit so that we have something to start of with. Be aware this is just the default project that is created when we create a new flutter project. As soon as we have something to go on. We will create a skeletal file structure and then bwork on or seperate branches and commit changes to the main branch as we go.