wrote some boilerplate code to create a generic feed display screen
Added a couple of different directories to estabilish a file directory layout.

Created multiple dart files for the podcast screen.

Currently the app only has 1 screen that displays lorem ipsums which is roughly in the format that we would like at the end.

To make things a little more clear.

The util folder holds the style.dart file which holds all the theme data.

assets folders has the fonts and images folder that are currently empty but that is where we shall place the fonts and images.

lib/screens folder has going to have a separate folder inside that pertains to a specific screen (layout) in the app.

each of these screen folders will include the files that include the code that directly relates to the screen in question nothing else.

for lib/screens/podcast_feed
this folder contains 2 files.
feed_tile.dart and
podcast_feed.dart
podcast_feed .dart file is the main layout of the podcast feed screen and is currently set as the home screen of the app.
the feed_tile.dart is a widget that that will be used to populate the podcast feed page. A tile holds all the information pertaining to a single podcast.

Hope this makes things a little more clearer.