PreMVC commit
Created the model to hold all the podcast data.

Webfeed library is used to get the RSSFeed and then the Feeds are turned into a a FeedData object, That holds the GUID, Title, URL and Description in it.

The FeedData list is shown as a scrollable list on the home screen.