Post MVC pattern implementation
Removed the file rss_data.dat since it is not needed anymore.

Implemented a MVC pattern to separate views from the Business logic.

All network calls are made through network_calls.dart

The screens directory is renamed to views to ensure homogeneity across naming conventions.

The views only speak to the controller. This is essential.

The widgets created for this  project are now moved from screens folder to custom_widgets folder in the lib directory.

The search method to filter the feed list is implemented. Currently we can only filter the list by querying the podcast title.